applications/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
azure-cosmos
python-dotenv